requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pillow==11.1.0
pytest==8.3.4
httpx==0.28.1
PyYAML==6.0.2
